techdemobackend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psutil
numpy
kubernetes
prometheus-client
prometheus-fastapi-instrumentator
psycopg2-binary
itsdangerous